requirements.txt
ccxt~=1.39.80
numpy~=1.19.4
pandas~=1.2.0
APScheduler~=3.6.3
python-dateutil~=2.8.1